refactor: seed the dashboard query from the parent server layout
Sonar failed the gate on new_duplicated_lines_density (3.6% vs 3%). The cause
was the split itself, not new duplicated code: renaming the client layout to
ConsultantDashboardShell.tsx makes Sonar treat all ~700 lines as new, and the
long-standing duplication between the consultant and consultee layouts then
counts as new duplication. get_duplications confirmed both blocks pair the
renamed file against consultee/[consulteeId]/layout.tsx.

So drop the rename entirely and seed from app/dashboard/layout.tsx, which is
already a server component that resolves the session. Strictly better:

- Both client layouts stay byte-identical to dev, so no rename, no duplication.
- One seed now covers consultant AND consultee, which share the same gate.
- The seed is keyed on the SESSION user, which this layout already has, rather
  than a profile id from the URL.

The nav-targets test goes back to dev's version too, since layout.tsx is once
again where the nav lives.
